+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## v0.5 - 2023-04-28
 
 - Allied reports now inform you when your team is the sole presence on a planet
 - Situation reports won't be issued by defeated allies